refactor: consolidate status history recording and popover shell
- Drop unused STATUS_HISTORY_TRIGGERS and _status_id/_status_name; add a
  single project_board_statuses() resolver feeding both status_name and
  _project_status_ids
- write_status_change now resolves status names internally from the project,
  so call sites pass only status ids and names cannot drift from ids
- Extract a shared AnchorPopover shell for the duplicated positioning and
  dismiss logic used by AssignmentChainPopover and StatusHistoryPopover
